[BEAM-10672] Move StatefulLoadGenerator from pardo_test.py
