+++ sdks/python/apache_beam/testing/synthetic_pipeline.py
@@ -41,11 +41,15 @@ import argparse
 import json
 import logging
 import math
+import os
 import sys
 import time
 from random import Random
+from typing import Tuple
 
 import apache_beam as beam
+from apache_beam import pvalue
+from apache_beam import typehints
 from apache_beam.io import WriteToText
 from apache_beam.io import iobase
 from apache_beam.io import range_trackers
@@ -55,6 +59,7 @@ from apache_beam.io.restriction_trackers import OffsetRestrictionTracker
 from apache_beam.options.pipeline_options import PipelineOptions
 from apache_beam.options.pipeline_options import SetupOptions
 from apache_beam.testing.test_pipeline import TestPipeline
+from apache_beam.transforms import userstate
 from apache_beam.transforms.core import RestrictionProvider
 
 try:
@@ -883,3 +888,67 @@ def run(argv=None, save_main_session=True):
 if __name__ == '__main__':
   logging.getLogger().setLevel(logging.INFO)
   run()
+
+
+class StatefulLoadGenerator(beam.PTransform):
+  """A PTransform for generating random data using Timers API."""
+  def __init__(self, input_options, num_keys=100):
+    self.num_records = input_options['num_records']
+    self.key_size = input_options['key_size']
+    self.value_size = input_options['value_size']
+    self.num_keys = num_keys
+
+  @typehints.with_output_types(Tuple[bytes, bytes])
+  class GenerateKeys(beam.DoFn):
+    def __init__(self, num_keys, key_size):
+      self.num_keys = num_keys
+      self.key_size = key_size
+
+    def process(self, impulse):
+      for _ in range(self.num_keys):
+        key = os.urandom(self.key_size)
+        yield key, b''
+
+  class GenerateLoad(beam.DoFn):
+    state_spec = userstate.CombiningValueStateSpec(
+        'bundles_remaining', combine_fn=sum)
+    timer_spec = userstate.TimerSpec('timer', userstate.TimeDomain.WATERMARK)
+
+    def __init__(self, num_records_per_key, value_size, bundle_size=1000):
+      self.num_records_per_key = num_records_per_key
+      self.payload = os.urandom(value_size)
+      self.bundle_size = bundle_size
+      self.time_fn = time.time
+
+    def process(
+        self,
+        _element,
+        records_remaining=beam.DoFn.StateParam(state_spec),
+        timer=beam.DoFn.TimerParam(timer_spec)):
+      records_remaining.add(self.num_records_per_key)
+      timer.set(0)
+
+    @userstate.on_timer(timer_spec)
+    def process_timer(
+        self,
+        key=beam.DoFn.KeyParam,
+        records_remaining=beam.DoFn.StateParam(state_spec),
+        timer=beam.DoFn.TimerParam(timer_spec)):
+      cur_bundle_size = min(self.bundle_size, records_remaining.read())
+      for _ in range(cur_bundle_size):
+        records_remaining.add(-1)
+        yield key, self.payload
+      if records_remaining.read() > 0:
+        timer.set(0)
+
+  def expand(self, pbegin):
+    assert isinstance(pbegin, pvalue.PBegin), (
+        'Input to transform must be a PBegin but found %s' % pbegin)
+    return (
+        pbegin
+        | 'Impulse' >> beam.Impulse()
+        | 'GenerateKeys' >> beam.ParDo(
+            StatefulLoadGenerator.GenerateKeys(self.num_keys, self.key_size))
+        | 'GenerateLoad' >> beam.ParDo(
+            StatefulLoadGenerator.GenerateLoad(
+                self.num_records // self.num_keys, self.value_size)))
